requirements.txt
pandas>=1.1.5
Pillow>=8.0.1
numpy==1.19.3
requests==2.25.1
nltk==3.5
PyDispatcher>=2.0.5
beautifulsoup4>=4.9.3
w3lib>=1.22.0
